Add fixture for pypi package sdbus 0.12.0
A fix has been implemented to correctly parse the LGPL license for certain pypi packages, which is reflected in pull request #518 in the crawler. The production definition was outdated prior to this fix.

The raw harvested result for the package displays the classifier in the registryData as "License :: OSI Approved :: GNU Lesser General Public License v2 or later (LGPLv2+)", which was previously parsed as GPL before the fix.

The current dev deployment, reflects the result of this fix: declared: 'GPL-2.0 AND LGPL-2.0-or-later AND LGPL-2.1-or-later":
-  "GPL-2.0" was detected by licensee
- "LGPL-2.0-or-later" was detected by clearlydefined and parsed from the classifier as GNU Lesser General Public License v2 or later (LGPLv2+)
- "LGPL-2.1-or-later" was detected by scancode.

Add the fixture to address this, prior to updating the pypi schema to allow for reharvest.